requirements.txt
asgiref==3.4.1
boto3==1.18.38
botocore==1.21.38
certifi==2021.5.30
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.6.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
